+++ src/drm/base-drm-protocol.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable no-unused-vars */
 import getLogger from '../utils/logger'
-import Error from '../utils/player-error'
+import Error from '../utils/error/player-error'
 import DrmSupport from './drm-support'
 import {DrmScheme} from './drm-scheme'
 
@@ -11,11 +11,11 @@ export default class BaseDrmProtocol implements IDrmProtocol {
   static DrmScheme = DrmScheme;
 
   static canPlayDrm(drmData: Array<Object>): boolean {
-    throw new Error(Error.Severity.CRITICAL, Error.ategory.PLAYER, Error.Code.RUNTIME_ERROR, {info: 'static canPlayDrm'});
+    throw new Error(Error.Severity.CRITICAL, Error.ategory.PLAYER, Error.Code.RUNTIME_ERROR_METHOD_NOT_IMPLEMENTED, 'static canPlayDrm');
   }
 
   static setDrmPlayback(...args): void {
-    throw new Error(Error.Severity.CRITICAL, Error.Category.PLAYER, Error.Code.RUNTIME_ERROR, {info: 'static setDrmPlayback'});
+    throw new Error(Error.Severity.CRITICAL, Error.Category.PLAYER, Error.Code.RUNTIME_ERROR_METHOD_NOT_IMPLEMENTED, 'static setDrmPlayback');
   }
 
   static destroy(): void {
