+++ src/plugin/base-plugin.js
@@ -3,9 +3,7 @@ import Player from '../player'
 import getLogger from '../utils/logger'
 import * as Utils from '../utils/util'
 import EventManager from '../event/event-manager'
-import {Error} from "../utils/player-error";
-
-'../utils/player-error'
+import Error from "../utils/error/player-error";
 import FakeEvent from '../event/fake-event'
 
 /** The BasePlugin responsible to implement the plugin interface.
@@ -70,7 +68,7 @@ export default class BasePlugin implements IPlugin {
    * @abstract
    */
   static isValid(): boolean {
-    throw new Error(Error.everity.CRITICAL, Error.Category.PLAYER, Error.Code.RUNTIME_ERROR, {data: 'isValid()'});
+    throw new Error(Error.Severity.CRITICAL, Error.Category.PLAYER, Error.Code.RUNTIME_ERROR_METHOD_NOT_IMPLEMENTED, 'isValid()');
   }
 
   /**
