+++ src/utils/error/category.js
@@ -1,0 +1,32 @@
+//@flow
+
+const Category: { [level: string]: number } = {
+  /** Errors from the network stack. */
+  'NETWORK': 1,
+
+  /** Errors parsing text streams. */
+  'TEXT': 2,
+
+  /** Errors parsing or processing audio or video streams. */
+  'MEDIA': 3,
+
+  /** Errors parsing the Manifest. */
+  'MANIFEST': 4,
+
+  /** Errors related to streaming. */
+  'STREAMING': 5,
+
+  /** Errors related to DRM. */
+  'DRM': 6,
+
+  /** Miscellaneous errors from the player. */
+  'PLAYER': 7,
+
+  /** Errors related to Ads. */
+  'ADS': 8,
+
+  /** Errors in the database storage (offline). */
+  'STORAGE': 9
+}
+
+export {Category}
